Add files via upload
ProofofConcept.bat added to showcase a brute-force type of multithreading with batch; literally starting a bunch of other CMD instances to run individually. This isn't very performant obviously, since it instantiates like a million CMD prompts (even without UI) that each have to create and write to a file. It is *technically* faster than the current UpdatedSubnetScanner.bat, though.

Other than that, I've done a few things to the scanner itself, namely adding a few more constants to make refactoring a little easier.

The 'Tests.bat' file is a WIP, and will be used as a central testing suite for the scanner and the cleanup utility once complete.